Refactor to use obsidian exports as constants
Obsidian passes its exports through binding functions, which then get
added to by esbuild when re-exported.  This makes it much more tedious
to trace plugin code in the debugger when a simple expression like
`x instanceof WorkspaceLeaf` steps through two layers of function
calls (that are unnececssary since Obsidian exports only constants,
not variables, and same for the re-exports).

So, by making a constant whose value is a shallow copy of
`require("obsidian")`, we can access concrete classes and functions
directly without extra layers of debugging hell.  (It also cuts
down on execution time, but 99.99% of the time that's irrelevant.)