Simplified API for building settings UI:
You can now use field() and group() at the top level of a showSettings()
method to create fields and groups, and still do chaining for everything
else.  You can also pass them a `settingsBuilder(element)` to make them
a child of an arbitrary element. (Meaning you can use them in views or
dialogs, not just the settings tab.)

All relevant classes now have a with(...features) method, for convenience
in chaining.  Callbacks receive the instance, and the method returns
the instance.  This is used make trackOpen() a function so it's
separately importable (and therefore tree-shakable for a smaller plugin
if you don't track opens or use field groups).

This also permits the development of other modular features for fields
and groups, such as hotkey editing, data bindings, group search, etc.