Switch savepoints from arrays to linked lists
This makes rollback an O(N) operation instead of N**2 when
there are lots of "subtask" savepoints.  It also saves
memory allocation when creating savepoints, and recycles
nodes to keep further allocations to a minimum.  Once the
app reaches a stable state, there should be zero new memory
allocations outside of the savepoint instances themselves.